fix: resolve API routing and Docker build issues
Frontend:
- Add runtime API URL detection with getApiUrl() function
- Configure Next.js standalone output for Docker production
- Update login page to use correct base path
- Add inventory, suppliers, movements, and critical stock pages
- Create reusable data table and form components

Backend:
- Fix Google OAuth id_token handling
- Update CORS configuration for base path
- Add ROOT_PATH configuration for Traefik routing
- Improve inventory API schemas

Infrastructure:
- Remove dev mode overrides (volumes, command) from docker-compose
- Configure PgBouncer for session mode
- Update Traefik labels for correct path routing

Fixes:
- 404 errors on API calls (incorrect base path)
- Docker container crash loop (missing standalone output)
- Stale code in container (dev mode volume mounts)
- Authentication flow with Google OAuth

+++ backend/app/schemas/inventory.py
@@ -1,7 +1,7 @@
 from datetime import date, datetime
 from enum import Enum
 from typing import Optional
-from decimal import Decimal
+# from decimal import Decimal
 
 from pydantic import BaseModel, Field, ConfigDict, field_validator
 
@@ -48,7 +48,7 @@ class WineBase(BaseModel):
     vintage: int = Field(..., ge=1900, le=2100)
     type: WineType = WineType.OTHER
     denomination: Optional[str] = Field(None, max_length=100)  # DOCG, DOC, IGT, etc.
-    price: Decimal = Field(..., gt=0)
+    price: float = Field(..., gt=0)
     quantity: int = Field(default=0, ge=0)
     threshold: Optional[int] = Field(default=10, ge=0)
     barcode: Optional[str] = Field(None, max_length=50)
@@ -69,7 +69,7 @@ class WineUpdate(BaseModel):
     vintage: Optional[int] = Field(None, ge=1900, le=2100)
     type: Optional[WineType] = None
     denomination: Optional[str] = Field(None, max_length=100)
-    price: Optional[Decimal] = Field(None, gt=0)
+    price: Optional[float] = Field(None, gt=0)
     quantity: Optional[int] = Field(None, ge=0)
     threshold: Optional[int] = Field(None, ge=0)
     barcode: Optional[str] = Field(None, max_length=50)
